Waveform Subsystem is a Website to generate, save and load (and delete) waveforms to (from) the Database and to send custom Waveforms or saved ones from the Database to the Agilent Waveform Generator. It can also draw the waveforms, stored in the Database.
File description: waveform.json
This file contains the html code from the interface website and the javascript code which communicates with the server script (hello.py)
For the communication with this python script WebSocket are implemented. (Faster than autobahn-wamp, because JSON serialization of the waveformdata, sent to the website for drawing, took very long).
Server commands sent to the website are using text messages and the transmission of the waveform data for drawing is using a binary stream.
Commands sent to the server (e.g. requesting a waveform to draw / sending a saved waveform, etc.) are text messages only.
Dygraph is used for drawing the waveforms.