Save run protocol separately in document.
Add ability to loop in measurements.

Code is evaluated at the beginning of a loop.

Use handlebars also for code in the loops...

rename "Name:" to "Function:".

Add option to perform a dry run.
This allows checking if parameters/loops work correctly.

Only stringify args if they are an object.